Set ALLOW_EXPERIMENT_REREGISTRATION to false in WebPack build.
Gutenberg introduced a system of sharing private APIs in
https://github.com/WordPress/gutenberg/pull/46131. One of the
safeguards is a check preventing the same module from opting-in
twice so that contributors cannot easily gain access by pretending
to be a core module.

That safeguard is only meant for WordPress core and not for
the released `@wordpress` packages. However, right now it is
opt-out and must be explicitly disabled by developers wanting to
install the `@wordpress` packages. Let's make it opt-out instead.

This commit opts-out from that check in WordPress core by setting
the ALLOW_EXPERIMENT_REREGISTRATION to false. Once it's merged,
the Gutenberg plugin should be adjusted to use `true` as the
default value.