Add support for base metrics
This updates the performance workflow to record base metrics from the 6.1.1 release on every commit in order to normalize the metrics being recorded against a consistent baseline on each run. We do this to reduce the amount of noise that is introduced in the data by changes in the workers that run these measurements.

After recording the initial metrics for the current commit, the docker environment is updated to version 6.1.1 via a WP CLI command and then the metrics are taken again. This time, by passing an optional `--prefix=base` argument so that the results are saved to files with a `base-` prefix for each test. All of the results are then parsed and sent to the dashboard API as part of the `log-results.js` script.