Users: Adjust the initialization of the `$duplicated_keys` array in `wp_salt()`.
This avoids an endless loop if `get_current_user_id()` is used in a callback attached to the `gettext` filter.

With the translated phrase moved into a separate assignment, the function succeeds in setting the static `$duplicated_keys` array once and no longer goes into this code section on subsequent calls.

Follow-up to [54249].

Props adityaarora010196, SergeyBiryukov.
Fixes #57121.

git-svn-id: https://develop.svn.wordpress.org/trunk@55433 602fd350-edb4-49c9-b593-d223f7449a82