get_merged_data: use empty theme.json object as base instead of core
The PR https://github.com/WordPress/wordpress-develop/pull/3441
introduced a bug by which the get_merged_data method is not cleared
in certain situations.

By doing (pseudo-code):

```php
function get_merged_data( $origin ) {
    $result = static::get_core_data();

    if ( 'block' === $origin ) { $result->merge( static::get_block_data() ); }
    if ( 'theme' === $origin ) { $result->merge( static::get_theme_data() ); }
    if ( 'custom' === $origin ) { $result->merge( static::get_custom_data() ); }

    return $result;
}
```

the base object ($result) is modifying the $core object directly,
and $core ends up storing the consolidated values instead of only the $core ones.

This is problematic because $core data is cached. Take, for example,
the following scenario:

```php
$data = get_merged_data( 'custom' );
$data = get_merged_data( 'theme' );
```

The expected output for $data is that it should not have data coming
from the 'custom' origin, however, it does.

The fix is reverting the change and use an empty object as base (pseudo-code):

```php
function get_merged_data( $origin ) {
    $result = new WP_Theme_JSON();

    $result->merge( static::get_core_data() );
    if ( 'block' === $origin ) { $result->merge( static::get_block_data() ); }
    if ( 'theme' === $origin ) { $result->merge( static::get_theme_data() ); }
    if ( 'custom' === $origin ) { $result->merge( static::get_custom_data() ); }

    return $result;
}
```