Build/Test Tools: Update third-party `welcome` Action.
This updates the third-party action used to post a welcome message to pull requests opened by first time contributors.

This release updates the action to use Node.js version 16 instead of 12, the latter of which support has been deprecated for in GitHub Action runners.

The action has also changed from `bukboo/welcome-action` to `wow-action/welcome`.

Merges [54651] to the 6.1 branch.
See #56793.

git-svn-id: https://develop.svn.wordpress.org/branches/6.1@55488 602fd350-edb4-49c9-b593-d223f7449a82