Administration: Update dashboard welcome panel colors, remove broken link
Remove the green background from the "dot" accent, to prevent low contrast for the dismiss button. Use menu colors for the background on alternate color schemes. Remove the "Edit styles" link as there is no longer a direct link to the global styles section of the site editor.

Follow-up to [55451].

Props sabernhardt, ryokuhi, laurlittle, richtabor.
Reviewed by hellofromTonya.
Merges [55574] to the 6.2 branch.
Fixes #57759.



git-svn-id: https://develop.svn.wordpress.org/branches/6.2@55576 602fd350-edb4-49c9-b593-d223f7449a82