Fix non-strictly checking issue on line 1014 of schema.php
I found a non-strictly checking issue on line 1014 of the schema.php file in WordPress. To fix this issue, I replaced the == operator with the === operator, which ensures that the values being compared are of the same type. This will help to prevent potential bugs and improve the overall security of the code. Please review and merge this pull request. Thank you!