HTML API: Add `has_self_closing_flag()` to Tag Processor
Trac ticket: [#58009](https://core.trac.wordpress.org/ticket/58009#ticket)

In this patch we're adding `has_self_closing_flag()` to the HTML Tag Processor.
This exposes whether a currently-matched tag contains the self-closing flag `/`.

This information is critical for the evolution of the HTML API in order
to track and parse HTML structure, specifically, knowing whether an
HTML foreign element is self-closing or not.