Fixing Multi-line PHP Snippet Embedding in wp-admin/network/user-new.php
I've noticed that the file wp-admin/network/user-new.php on line 103 has a multi-line PHP snippet that violates the recommended coding standard of WordPress. Specifically, the closing PHP tag is not on a new line, making the code harder to read and understand.

To adhere to the WordPress coding standard and make the code easier to maintain, I propose to fix this issue by moving the closing PHP tag to a new line. I have created a pull request that implements this change.

Please review and merge the pull request if you find it acceptable. Thank you for your attention to this matter.