Fix "else if" usage in favor of "elseif" in WordPress coding standard
This pull request fixes the issue of using "else if" instead of "elseif" in the WordPress coding standard. According to the standard, "elseif" should be used instead of "else if" because the latter is not compatible with the colon syntax for if|elseif blocks.

I have identified three files where this issue exists and have made the necessary changes to replace all occurrences of "else if" with "elseif".