HTML API: Delete all copies of an attribute when removing and duplicates exist. (#4317)
When encountering an HTML tag with duplicate copies of an attribute the tag
processor ignores the duplicate values, according to the specification. However,
when removing an attribute it must remove all copies of that attribute lest
one of the duplicates becomes the primary and it appears as if no attributes
were removed.

In this patch we're adding tests that will be used to ensure that all attribute
copies are removed from a tag when one is requested to be removed.

Co-authored-by: Bernie Reiter <96308+ockham@users.noreply.github.com>