HTML API: Ensure attribute updates happen only once for case variants
When setting a new value for an attribute multiple times and providing
multiple case variations of the attribute name the Tag Processor has
been appending multiple copies of the attribute into the updated HTML.

This means that only the first attribute set determines the value in
the final output.

In this patch we're adding a test to catch the situation.