HTML API: Explore deferring attribute name allocations until needed
Inspired by a C port of the Tag Processor I wanted to experiment with the idea
of deferring the allocation of strings for the parsed attribute names until and
unless they are needed.

Effectively this means storing only the starting and ending string index at which
the names are found instead of storing the names themselves as strings.

While not all tests are passing in this patch, the idea was built soundly and
those tests are solvable with a little more code.

Unfortunately the Tag Processor appears to run around 4% slower in this branch
withs its indices than in the `trunk` branch with its allocations. While not
obvbious why it's slower that's a non-starter for the optimization.

To try and investigate further, I added a static allocation for the "class"
attribute under the idea that associative arrays might be the reason for the
slowness. So I created `$this->attr_class` which should make for near instant
lookup of the "class" attribute, but it didn't impact speed at all.

I have thought that if we can trap the most common set of attributes as static
properties then we might be able to avoid arrays in most tags. That _may_ have
a measurable speedup, but right now with only `class` it's non-conclusive.