HTML API: Fix ordering issue in docblock that's breaking Developer Resources
In this patch we're re-ordering the `@see` tag in the docblock comment for
`WP_HTML_Tag_Processor::get_attribute_names_with_prefix`. Previously the
rendering in Developer Resources was broken, whereby it merged the link
in the `@see` tag with the example code.

Now, in this patch, the `@see` tag has been moved below the code example so
that the example is properly extracted.

Additionally the example code for `WP_HTML_Tag_Processor::get_tag` has been
updated to show lowercase tag names in the input HTML so that it doesn't
convey the wrong impression that the uppercase output from `get_tag()`
depends on the case of the input HTML.