HTML API: Restore mistakenly-removed content in documentation.
In a3ed152c54c117d39d32a72a3d22f082dd5ffa7f the Unicode replacement character
was mistakenly removed, being believed to be a text rendering mistake, because
the character _is_ the replacement character used in situations where those
text mistakes occur.

However, the purpose of including the character was to communicate what it
looks like and why the Tag Processor won't insert it into the document (which
would be wrong from the Unicode point of view for such a streaming processor
that doesn't produce the final rendered view of the text).

In this patch we're bringing back the character and adding a small clue to
try and remove the confusion that previously led to its unwanted removal.